Backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httptools==0.6.1
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
outcome==1.3.0.post0
packaging==24.0
passlib==1.7.4
priority==2.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pymssql==2.2.11
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.36.3
trio==0.25.0
typing_extensions==4.10.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0

tenacity==8.2.3